requirements.txt
feedparser
google-generativeai
python-dotenv
pandas
python-dateutil